functions/requirements.txt
blinker==1.8.2
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudevents==1.9.0
colorclass==2.2.2
coverage==7.5.4
cryptography==42.0.8
deprecation==2.1.0
docopt==0.6.2
firebase-admin==6.5.0
firebase-functions==0.4.2
Flask==3.0.3
Flask-Cors==4.0.1
functions-framework==3.8.0
google-api-core==2.19.1
google-api-python-client==2.126.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.1
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-events==0.12.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
grpcio==1.64.1
grpcio-status==1.62.2
gunicorn==22.0.0
httplib2==0.22.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
msgpack==1.0.8
packaging==24.1
pip-upgrader==1.4.15
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyparsing==3.1.2
PyYAML==6.0.1
requests==2.32.3
rsa==4.9
terminaltables==3.1.10
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
watchdog==4.0.1
Werkzeug==3.0.3
